requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn 
uvicorn
whitenoise 
pillow